requirements.txt
monarchmoney==0.1.11
selenium
webdriver-manager
python-dotenv